feat: Add tray icon functionality
